XRDDEV-237 refactoring
